pyproject.toml
[project]
name = "entra-protocol-lab"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
  "flask==3.0.3",
  "Authlib==1.3.1",
  "requests>=2.32,<3",
  "Flask-Session>=0.5.0,<0.7",
  "python-dotenv>=1.0.1,<2",
  "gunicorn>=21,<23",
  "pysaml2>=7.5.4",
  "pyjwt>=2.10.1",
  "cryptography>=43.0.3",
]
